% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_route.R
\name{gh_route}
\alias{gh_route}
\title{Send Routing request to Graph Hopper}
\usage{
gh_route(Start, End, min_plateau_factor = 0.2, max_paths = 5,
  max_share_factor = 0.6)
}
\arguments{
\item{Start}{Start location, as a string, formatted as "lat,long"}

\item{End}{End location, as a string, formatted as "lat,long"}

\item{min_plateau_factor}{As per Graphhopper documentation}

\item{max_paths}{As per Graphhopper documentation}

\item{max_share_factor}{As per Graphhopper documentation}
}
\value{
A list containing dataframes with lat & long for every point along the route.
}
\description{
This function will send a request for alternative routes, and return a list
containing the data of the points along the route.
}
