% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_routes.R
\name{plot_routes}
\alias{plot_routes}
\title{Plot all generated routes}
\usage{
plot_routes(Routes)
}
\arguments{
\item{Routes}{Input List, containing dataframes with two columns: "longitude" and "latitude" and points ordered correctly.}
}
\value{
A leafleft map that will show the routes as indicated.
}
\description{
This function take as an input a list of dataframes 
containing the data of the points along a route, and 
produce a map in leaflet that visualises these as lines.
}
