% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_routes.R
\name{plot_routes}
\alias{plot_routes}
\title{Plot all generated routes}
\usage{
plot_routes(Routes)
}
\arguments{
\item{Routes}{Start location, as a string, formatted as "lat,long"}
}
\value{
A list containing dataframes with lat & long for every point along the route.
}
\description{
This function take as an input a list of dataframes 
containing the data of the points along a route, and 
produce a map in leaflet that visualises these as lines.
}
